Code for ESP with button and for ESP with a sensor
Problem of "no communication" between ESPs is solved. I assume that problem was weak power supply to ESP_softap. Or that ESP with button, that is in wifi_ap_sta mode, must start both AP and STA in the same channel.
Also, web browser works well by showing not only real traffic light lights, but also the special simulated lights, which depend on the reading from a sensor.